requirements.txt
azure-ai-formrecognizer==3.3.3
azure-ai-documentintelligence==1.0.2
httpx[http2]==0.28.1
fastapi==0.108.0
uvicorn==0.25.0
torch
torchaudio
transformers
tqdm==4.67.1
python-dotenv==1.0.1
python-multipart==0.0.20
pandas
aiofiles==24.1.0
pyannote-audio
pyannote-core
pyannote-database
pyannote-metrics
pyannote-pipeline
openai-whisper==20250625
websockets
librosa==0.11.0
soundfile==0.13.1
pydub==0.25.1